Rama actualizacion (#53)
* Actualizando mi rama de actualizacion (#52)

* Actulalizada al ingles la plantilla con favoritos y categorias

* Mas vistas con ingles puesto (#50)

---------


* Solucionando errores

He solucionado el error de que te deje añadir un producto repetido al carrito. Que el producto que se añade al carrito sea el que corresponde. Modificado el migrate para que los usuarios se le añada es en el idioma.
Ademas solucionando un campo faltante de un input  y un fallo que habia si creabamos un producto sin categoria.

* Añadida nueva funcionalidad en el carrito

Ahora podemos aumentar y disminuir la cantidad en el carrito.

---------

+++ CarUPO/app/Http/Controllers/Carrito_comprasController.php
@@ -13,6 +13,7 @@ use App\Models\Producto;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Mail;
+use Illuminate\Support\Facades\DB;
 
 
 class Carrito_comprasController extends Controller
@@ -27,30 +28,62 @@ class Carrito_comprasController extends Controller
 
 
 
-    public function eliminarLineaCarrito(Request $request)
+    public function aumentarLineaCarrito(Request $request)
+    {
+
+        $id = Auth::user()->id;
+        $mi_carrito = Carrito_compra::findOrFail($id);
+        $linea_carrito = Linea_carrito::findOrFail($request->id);
+
+        $linea_carrito->cantidad +=  1;
+        $mi_carrito->precio_total += $linea_carrito->producto->precio;
+        $linea_carrito->precio_parcial += $linea_carrito->producto->precio;
+        $linea_carrito->save();
+        $mi_carrito->save();
+
+        return app()->make(Carrito_comprasController::class)->callAction('mostrarCarrito', []);
+    }
+
+    public function disminuirLineaCarrito(Request $request)
     {
         $id = Auth::user()->id;
         $mi_carrito = Carrito_compra::findOrFail($id);
         $linea_carrito = Linea_carrito::findOrFail($request->id);
-        $mi_carrito->precio_total = $mi_carrito->precio_total - $linea_carrito->precio_parcial;
+
+
+        $linea_carrito->cantidad -=  1;
+        $mi_carrito->precio_total -= $linea_carrito->producto->precio;
+        $linea_carrito->precio_parcial -= $linea_carrito->producto->precio;
+        $linea_carrito->save();
         $mi_carrito->save();
-        $linea_carrito->delete();
+
+        if ($mi_carrito->precio_total == 0) {
+            $linea_carrito->delete();
+        }
         return app()->make(Carrito_comprasController::class)->callAction('mostrarCarrito', []);
     }
-
     public function addToCarrito(Request $request)
     {
         $id = Auth::user()->id;
-        $mi_carrito = Carrito_compra::findOrFail($id);
-        $linea_carrito = new Linea_carrito();
-        $producto = Producto::findOrFail($id);
-        $linea_carrito->fk_producto_id = $producto->id;
-        $linea_carrito->fk_carrito_id = $mi_carrito->id;
-        $linea_carrito->cantidad = $request->cantidad;
-        $linea_carrito->precio_parcial = $request->cantidad * $producto->precio;
-        $linea_carrito->save();
-        $mi_carrito->precio_total = $mi_carrito->precio_total + $linea_carrito->precio_parcial;
-        $mi_carrito->save();
+        $producto = Producto::findOrFail($request->id);
+
+        $existe_producto_en_carrito = DB::table('carrito_compras')
+            ->join('linea_carritos', 'carrito_compras.id', '=', 'linea_carritos.fk_carrito_id')
+            ->where('linea_carritos.fk_producto_id', '=', $producto->id)
+            ->where('carrito_compras.fk_user', '=', Auth::id())
+            ->exists();
+
+        if (!$existe_producto_en_carrito) {
+            $mi_carrito = Carrito_compra::findOrFail($id);
+            $linea_carrito = new Linea_carrito();
+            $linea_carrito->fk_producto_id = $producto->id;
+            $linea_carrito->fk_carrito_id = $mi_carrito->id;
+            $linea_carrito->cantidad = $request->cantidad;
+            $linea_carrito->precio_parcial = $request->cantidad * $producto->precio;
+            $linea_carrito->save();
+            $mi_carrito->precio_total = $mi_carrito->precio_total + $linea_carrito->precio_parcial;
+            $mi_carrito->save();
+        }
         return app()->make(Carrito_comprasController::class)->callAction('mostrarCarrito', []);
     }
